feat: Basic diary app
