--- accounts/templates/accounts/register.html
@@ -1,35 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.6.1/dist/css/bootstrap.min.css" integrity="sha384-zCbKRCUGaJDkqS1kPbPd7TveP5iyJE0EjAuZQTgFLD2ylzuqKfdKlfG/eSrtxUkn" crossorigin="anonymous">
-
-    <title>Document</title>
-</head>
-<body class="m-5">
-<h1 class="display-1 my-5">Register</h1>
-<form method="post">
-    {% csrf_token %}
-    {{ form.as_p }}
-    {#    <div class="form-group">#}
-    {#        <label for="email">Email</label>#}
-    {#        {{ form.username }}#}
-    {##}
-    {#    </div>#}
-    {#    <div class="form-group">#}
-    {#        <label for="email">Password</label>#}
-    {#        {{ form.password1 }}#}
-    {#    </div>#}
-    {#    <div class="form-group">#}
-    {#        <label for="email">Password Confirmation</label>#}
-    {#        {{ form.password2 }}#}
-    {#    </div>#}
-    <input class="btn btn-primary" type="submit" value="Register">
-</form>
-<p>
-    Already a User? <a href="/accounts/login/">Signin</a>
-</p>
-</body>
-</html>
