--- accounts/templates/registration/login.html
@@ -1,36 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.6.1/dist/css/bootstrap.min.css" integrity="sha384-zCbKRCUGaJDkqS1kPbPd7TveP5iyJE0EjAuZQTgFLD2ylzuqKfdKlfG/eSrtxUkn" crossorigin="anonymous">
-    <title>Document</title>
-</head>
-<body class="m-5">
-<h1 class="display-1 mb-5">Login</h1>
-<form method="post" action="{% url 'login' %}">
-    {% csrf_token %}
-    <div class="form-group">
-        <label for="{{ form.username.id_for_label }}">Username : </label>
-        {{ form.username }}
-    </div>
-    <div class="form-group">
-        <label for="{{ form.password.id_for_label }}">Password : </label>
-        {{ form.password }}
-    </div>
-    {% if form.errors %}
-        <p class=" label label-danger text-danger">
-            Your username and password didn't match.
-            Please try again.
-        </p>
-    {% endif %}
-    <input class="btn btn-primary" type="submit" value="Login">
-
-</form>
-<br>
-<p>
-    New User? <a href="/accounts/register/">Create an Account</a>
-</p>
-</body>
-</html>
