--- accounts/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-
-from accounts import views
-
-urlpatterns = [
-    path('register/', views.register_view, name='register')
-]
