+++ diaries/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import Diary
+# Register your models here.
+
+admin.site.register(Diary)
