--- diaryApp/forms.py
@@ -1,12 +1,0 @@
-from django.contrib.auth import get_user_model
-from django.contrib.auth.forms import UserCreationForm
-
-
-class UserAdminCreationForm(UserCreationForm):
-    """
-    A Custom form for creating new users.
-    """
-
-    class Meta:
-        model = get_user_model()
-        fields = ['phone']
