--- diaryApp/static/css/style.css
@@ -1,241 +1,0 @@
-/* Add a black background color to the top navigation bar */
-/*.topnav {*/
-/*    overflow: hidden;*/
-/*    background-color: #e9e9e9;*/
-/*}*/
-
-/*!* Style the links inside the navigation bar *!*/
-/*.topnav a {*/
-/*    float: left;*/
-/*    display: block;*/
-/*    color: black;*/
-/*    text-align: center;*/
-/*    padding: 14px 16px;*/
-/*    text-decoration: none;*/
-/*    font-size: 17px;*/
-/*}*/
-
-/*!* Change the color of links on hover *!*/
-/*.topnav a:hover {*/
-/*    background-color: #ddd;*/
-/*    color: black;*/
-/*}*/
-
-/*!* Style the "active" element to highlight the current page *!*/
-/*.topnav a.active {*/
-/*    background-color: #597e9c;*/
-/*    color: white;*/
-/*}*/
-
-/*!* Style the search box inside the navigation bar *!*/
-/*.topnav input[type=text] {*/
-/*    float: right;*/
-/*    padding: 6px;*/
-/*    border: none;*/
-/*    margin-top: 8px;*/
-/*    margin-right: 16px;*/
-/*    font-size: 17px;*/
-/*}*/
-
-/*!* When the screen is less than 600px wide, stack the links and the search field vertically instead of horizontally *!*/
-/*@media screen and (max-width: 600px) {*/
-
-/*    .topnav a,*/
-/*    .topnav input[type=text] {*/
-/*        float: none;*/
-/*        display: block;*/
-/*        text-align: center;*/
-/*        width: 100%;*/
-/*        margin: 0;*/
-/*        padding: 14px;*/
-/*    }*/
-
-/*    .topnav input[type=text] {*/
-/*        border: 1px solid #ccc;*/
-/*    }*/
-/*}*/
-
-/*h2 {*/
-/*    text-align: center;*/
-/*}*/
-
-/*h4 {*/
-/*    text-align: center;*/
-/*}*/
-
-/*form {*/
-/*    padding-left: 9rem;*/
-/*    padding-right: 9rem;*/
-/*    padding-bottom: 5%;*/
-/*    border: 3px solid palegoldenrod;*/
-/*}*/
-
-
-/*input[type=text],*/
-/*input[type=password] {*/
-/*    width: 100%;*/
-/*    padding: 12px 20px;*/
-/*    margin: 8px 0;*/
-/*    display: inline-block;*/
-/*    border: 3px solid #ccc;*/
-/*    box-sizing: border-box;*/
-/*}*/
-
-
-/*button {*/
-/*    background-color: grey;*/
-/*    color: white;*/
-/*    padding: 14px 20px;*/
-/*    margin: 8px 0;*/
-/*    border: none;*/
-/*    cursor: pointer;*/
-/*    width: 100%;*/
-/*}*/
-
-
-/*button:hover {*/
-/*    opacity: 0.8;*/
-/*}*/
-
-
-/*.cancelbtn {*/
-/*    width: auto;*/
-/*    padding: 10px 18px;*/
-/*    background-color: #f44336;*/
-
-/*}*/
-
-/*.register_new_user {*/
-/*    width: auto;*/
-/*    padding: 10px 18px;*/
-/*    background-color: cornflowerblue;*/
-/*}*/
-
-/*.imgcontainer {*/
-/*    text-align: center;*/
-/*    margin: 24px 0 12px 0;*/
-/*}*/
-
-/*.imgcontainer img {*/
-/*    border-radius: 100%;*/
-/*    height: 300px;*/
-/*    width: 300px;*/
-/*}*/
-
-
-/*img.avatar {*/
-/*    width: 40%;*/
-/*    height: 50%;*/
-/*    border-radius: 50%;*/
-/*}*/
-
-
-/*.container {*/
-/*    padding: 16px;*/
-/*}*/
-
-
-/*span.psw {*/
-/*    float: right;*/
-/*    padding-top: 16px;*/
-/*}*/
-
-
-/*@media screen and (max-width: 300px) {*/
-/*    span.psw {*/
-/*        display: block;*/
-/*        float: none;*/
-/*    }*/
-
-/*    .cancelbtn {*/
-/*        width: 100%;*/
-/*    }*/
-/*}*/
-
-/** {*/
-/*    box-sizing: border-box*/
-/*}*/
-
-/*!* Full-width input fields *!*/
-/*input[type=text],*/
-/*input[type=password],*/
-/*input[type=number] {*/
-/*    width: 100%;*/
-/*    padding: 15px;*/
-/*    margin: 5px 0 22px 0;*/
-/*    display: inline-block;*/
-/*    border: none;*/
-/*    background: #f1f1f1;*/
-/*}*/
-
-/*input[type=text]:focus,*/
-/*input[type=password]:focus {*/
-/*    background-color: #ddd;*/
-/*    outline: none;*/
-/*}*/
-
-/*hr {*/
-/*    border: 1px solid burlywood;*/
-/*    margin-bottom: 25px;*/
-/*}*/
-
-/*!* Set a style for all buttons *!*/
-/*button {*/
-/*    background-color: grey;*/
-/*    color: white;*/
-/*    padding: 14px 20px;*/
-/*    margin: 8px 0;*/
-/*    border: none;*/
-/*    cursor: pointer;*/
-/*    width: 100%;*/
-/*    opacity: 0.9;*/
-/*}*/
-
-/*button:hover {*/
-/*    opacity: 1;*/
-/*}*/
-
-/*!* Extra styles for the cancel button *!*/
-/*.cancelbtn {*/
-/*    padding: 14px 20px;*/
-/*    background-color: #f44336;*/
-/*}*/
-
-/*!* Float cancel and signup buttons and add an equal width *!*/
-/*.cancelbtn,*/
-/*.signupbtn {*/
-/*    float: left;*/
-/*    width: 50%;*/
-/*}*/
-
-/*!* Add padding to container elements *!*/
-/*.container {*/
-/*    padding: 16px;*/
-/*}*/
-
-/*h1 {*/
-/*    text-align: center;*/
-/*}*/
-
-/*!* Clear floats *!*/
-/*.clearfix::after {*/
-/*    content: "";*/
-/*    clear: both;*/
-/*    display: table;*/
-/*}*/
-
-/*!* Change styles for cancel button and signup button on extra small screens *!*/
-/*@media screen and (max-width: 300px) {*/
-
-/*    .cancelbtn,*/
-/*    .signupbtn {*/
-/*        width: 100%;*/
-/*    }*/
-/*}*/
-
-@media all and (max-width: 480px) {
-    .custom-class {
-        width: 100%;
-        display: block;
-    }
-}
