--- diaryApp/templates/diary.html
@@ -1,38 +1,0 @@
-{% extends "base.html" %}
-{% load diary_extras %}
-{% block content %}
-
-
-    <nav class="navbar navbar-expand-lg navbar-light bg-light">
-        <a class="navbar-brand" href="/">Daybook</a>
-        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-            <span class="navbar-toggler-icon"></span>
-        </button>
-
-        <div class="collapse navbar-collapse" id="navbarSupportedContent">
-            <a class="ml-auto mr-5" href="/">
-                <button class="btn btn-outline-danger btn-block  mx-md-3 my-md-2"> Exit and Go Home</button>
-            </a>
-        </div>
-    </nav>
-    <div class="container my-5">
-
-        <h1 class="display-1">{{ data.title|convert_markdown|safe }}</h1>
-        <p>{{ data.content|convert_markdown|safe }}</p>
-        <p>{{ data.date }}</p>
-        <br>
-        <hr>
-        <br>
-        <a href="/edit_diary/{{ data.id }}">
-            <button class="btn btn-outline-primary btn-block" type="submit">
-                Edit
-            </button>
-        </a>
-        <br>
-        <a href="/delete_diary/{{ data.id }}">
-            <button class="btn btn-outline-danger btn-block" type="submit">
-                Delete
-            </button>
-        </a>
-    </div>
-{% endblock %}
