--- diaryApp/templates/edit_diary.html
@@ -1,33 +1,0 @@
-{% extends "base.html" %}
-{% load diary_extras %}
-{% block content %}
-
-
-    <nav class="navbar navbar-expand-lg navbar-light bg-light">
-        <a class="navbar-brand" href="/">Daybook</a>
-        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-            <span class="navbar-toggler-icon"></span>
-        </button>
-
-        <div class="collapse navbar-collapse" id="navbarSupportedContent">
-            <a class="ml-auto mr-5" href="/">
-                <button class="btn btn-outline-danger btn-block  mx-md-3 my-md-2"> Exit and Go Home</button>
-            </a>
-        </div>
-    </nav>
-    <div class="container my-5">
-        <form method="post" action="/edit_diary_submit/{{ diary.id }}">
-            {% csrf_token %}
-            <div class="form-group">
-                <label for="Title">Title</label>
-                <input type="text" value="{{ diary.title }}" name="Title" class="form-control" id="Title"
-                       aria-describedby="emailHelp">
-            </div>
-            <div class="form-group">
-                <label for="Content">Content</label>
-                <textarea class="form-control" name="Content" id="Content" rows="20">{{ diary.content }}</textarea>
-            </div>
-            <button type="submit" class="btn btn-success btn-block btn-lg">Submit</button>
-        </form>
-    </div>
-{% endblock %}
