--- diaryApp/templates/entry_page.html
@@ -1,31 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-
-    <nav class="navbar navbar-expand-lg navbar-light bg-light">
-        <a class="navbar-brand" href="/">Daybook</a>
-        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-            <span class="navbar-toggler-icon"></span>
-        </button>
-
-        <div class="collapse navbar-collapse" id="navbarSupportedContent">
-            <a class="ml-auto mr-5" href="/">
-                <button class="btn btn-outline-danger btn-block  mx-md-3 my-md-2"> Exit and Go Home</button>
-            </a>
-        </div>
-    </nav>
-    <div class="container my-2">
-
-        <form action="/new_diary/" method="post">
-            {% csrf_token %}
-            <div class="form-group">
-                <label for="title">Title</label>
-                <input type="text" name="title" class="form-control" id="title" aria-describedby="emailHelp">
-            </div>
-            <div class="form-group">
-                <label for="content">Diary Body</label>
-                <textarea class="form-control" name="content" id="content" rows="20"></textarea>
-            </div>
-            <button type="submit" class="btn btn-primary btn-block">Submit</button>
-        </form>
-    </div>
-{% endblock %}
