--- diaryApp/templates/index.html
@@ -1,40 +1,0 @@
-{% extends 'base.html' %}
-{% load diary_extras %}
-{% block content %}
-
-    <nav class="navbar navbar-expand-lg navbar-light bg-light">
-        <a class="navbar-brand display-1" href="/"><h1>DayBook</h1></a>
-        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent"
-                aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-            <span class="navbar-toggler-icon"></span>
-        </button>
-
-        <div class="collapse navbar-collapse" id="navbarSupportedContent">
-            <ul class="navbar-nav mr-auto">
-            </ul>
-            <a href="/new_entry/">
-                <button class="btn btn-outline-primary mx-md-3 my-2 my-sm-0 custom-class" type="submit">
-                    New Diary
-                </button>
-            </a><a href="/accounts/logout/">
-            <button class="btn btn-outline-danger btn-sm mx-md-3 my-2 my-sm-0 custom-class" type="submit">
-                Sign Out
-            </button>
-        </a>
-        </div>
-    </nav>
-    <div class="container">
-        <h1 class="display-1 mb-3">Memories</h1>
-        {% for d in data %}
-            <div class="card my-2">
-                <div class="card-body">
-                    <h5 class="card-title">{{ d.title|convert_markdown|safe }}</h5>
-                    <p>{{ d.date }}</p>
-                    <a href="/view_diary/{{ d.id }}" class="btn btn-primary btn-block   my-2 my-sm-0
-                            custom-class">Read More</a>
-                </div>
-            </div>
-        {% endfor %}
-
-    </div>
-{% endblock %}
