--- diaryApp/templates/login.html
@@ -1,27 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-
-    <div class="container my-5">
-        <h1 class="text-center mb-5 display-1">Login</h1>
-        <form method="post" action="/do_login/">
-            {% csrf_token %}
-            <div class="form-group">
-                <label for="email">Username</label>
-                <input type="email" name="email" class="form-control" id="email" aria-describedby="emailHelp">
-                <small id="emailHelp" class="form-text text-muted">Your email is your username.</small>
-            </div>
-            <div class="form-group">
-                <label for="password">Password</label>
-                <input type="password" class="form-control" id="password" name="password">
-            </div>
-            <button type="submit" class="btn btn-dark btn-block btn-lg">Submit</button>
-        </form>
-        <br>
-        <hr>
-        <br>
-        <p>Didn't registered yet?</p>
-        <a href="/register/">
-            <button type="submit" class="btn btn-outline-dark btn-block btn-lg">Create an Account</button>
-        </a>
-    </div>
-{% endblock %}
