--- diaryApp/templates/register.html
@@ -1,34 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-    <div class="container my-5">
-        <h1 class="text-center mb-5 display-1">Register</h1>
-        <form method="post" action="/new_register/">
-            {% csrf_token %}
-            <div class="form-group">
-                <label for="fullname">Full Name</label>
-                <input type="text" class="form-control" id="fullname" name="fullname" aria-describedby="fullname">
-            </div>
-            <div class="form-group">
-                <label for="email">Email</label>
-                <input type="email" name="email" class="form-control" id="email" aria-describedby="emailHelp">
-                <small id="emailHelp" class="form-text text-muted">Your email will be your username.</small>
-            </div>
-            <div class="form-group">
-                <label for="password">Password</label>
-                <input type="password" name="password" class="form-control" id="password">
-            </div>
-            <div class="form-group">
-                <label for="password1">Confirm Password</label>
-                <input type="password" name="password1" class="form-control" id="password1">
-            </div>
-            <button type="submit" class="btn btn-dark btn-block btn-lg">Submit</button>
-        </form>
-        <br>
-        <hr>
-        <br>
-        <p>Already a Customer?</p>
-        <a href="/login/">
-            <button type="submit" class="btn btn-outline-dark btn-block btn-lg">Login</button>
-        </a>
-    </div>
-{% endblock %}
