+++ requirements.txt
@@ -1,0 +1,7 @@
+asgiref==3.5.0
+autopep8==1.6.0
+backports.zoneinfo==0.2.1
+Django==4.0.2
+pycodestyle==2.8.0
+sqlparse==0.4.2
+toml==0.10.2
