+++ templates/navbar.html
@@ -1,0 +1,10 @@
+<header>
+    <h1>Diary App</h1>
+    {% if user.is_authenticated %}
+        Logged in as : {{ user }}
+        <a href="{% url 'logout' %}">Logout</a>
+    {% else %}
+        <a href="{% url 'login' %}">Login</a>
+        <a href="{% url 'signup' %}">Signup</a>
+    {% endif %}
+</header>
